Rename the differentiation fitness function. Refactoring and extraction of fitness predictor thread
