+++ src/main/resources/logback.xml
@@ -8,6 +8,7 @@
     </appender>
 
     <logger name="pl.edu.agh.hubert.languages.math.MathRandomIndividualGenerator" level="INFO"/>
+    <logger name="pl.edu.agh.hubert.engine.DeterministicCrowdingEvolutionEngine" level="DEBUG"/>
 
     <root level="debug">
         <appender-ref ref="STDOUT" />
